+++ src/biz/bokhorst/xprivacy/ApplicationInfoEx.java
@@ -17,6 +17,9 @@ import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.PackageManager.NameNotFoundException;
+import android.content.res.Resources;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
@@ -31,7 +34,6 @@ public class ApplicationInfoEx implements Comparable<ApplicationInfoEx> {
 	private Map<String, PackageInfo> mMapPkgInfo = new HashMap<String, PackageInfo>();
 
 	// Cache
-	private Drawable mIcon = null;
 	private Boolean mInternet = null;
 	private Boolean mFrozen = null;
 	private long mInstallTime = -1;
@@ -149,11 +151,35 @@ public class ApplicationInfoEx implements Comparable<ApplicationInfoEx> {
 	}
 
 	public Drawable getIcon(Context context) {
-		if (mIcon == null)
-			// Pick first icon
-			if (mMapAppInfo.size() > 0)
-				mIcon = mMapAppInfo.firstEntry().getValue().loadIcon(context.getPackageManager());
-		return (mIcon == null ? new ColorDrawable(Color.TRANSPARENT) : mIcon);
+		// Pick first icon
+		if (mMapAppInfo.size() > 0)
+			return mMapAppInfo.firstEntry().getValue().loadIcon(context.getPackageManager());
+		else
+			return new ColorDrawable(Color.TRANSPARENT);
+	}
+
+	public Bitmap getIconBitmap(Context context) {
+		if (mMapAppInfo.size() > 0) {
+			try {
+				final ApplicationInfo appInfo = mMapAppInfo.firstEntry().getValue();
+				if (appInfo.icon == 0)
+					appInfo.icon = android.R.drawable.sym_def_app_icon;
+				final Resources resources = context.getPackageManager().getResourcesForApplication(appInfo);
+
+				final BitmapFactory.Options options = new BitmapFactory.Options();
+				options.inJustDecodeBounds = true;
+				BitmapFactory.decodeResource(resources, appInfo.icon, options);
+
+				final int pixels = Math.round(Util.dipToPixels(context, 48));
+				options.inSampleSize = Util.calculateInSampleSize(options, pixels, pixels);
+				options.inJustDecodeBounds = false;
+				return BitmapFactory.decodeResource(resources, appInfo.icon, options);
+			} catch (NameNotFoundException ex) {
+				Util.bug(null, ex);
+				return null;
+			}
+		} else
+			return null;
 	}
 
 	public boolean hasInternet(Context context) {