+++ src/biz/bokhorst/xprivacy/Util.java
@@ -12,6 +12,7 @@ import java.io.OutputStream;
 import java.lang.StackOverflowError;
 import java.io.UnsupportedEncodingException;
 import java.lang.reflect.Method;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.RuntimeException;
 import java.net.SocketTimeoutException;
 import java.net.UnknownHostException;
@@ -37,6 +38,7 @@ import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.PackageManager.NameNotFoundException;
+import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Environment;
@@ -45,7 +47,9 @@ import android.os.RemoteException;
 import android.os.TransactionTooLargeException;
 import android.os.UserHandle;
 import android.util.Base64;
+import android.util.DisplayMetrics;
 import android.util.Log;
+import android.util.TypedValue;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
@@ -101,6 +105,12 @@ public class Util {
 	}
 
 	public static void bug(XHook hook, Throwable ex) {
+		if (ex instanceof InvocationTargetException) {
+			InvocationTargetException exex = (InvocationTargetException) ex;
+			if (exex.getTargetException() != null)
+				ex = exex.getTargetException();
+		}
+
 		int priority;
 		if (ex instanceof ActivityShare.AbortException)
 			priority = Log.WARN;
@@ -599,5 +609,25 @@ public class Util {
 				views.add(child);
 		}
 		return views;
+	}
+
+	public static float dipToPixels(Context context, float dipValue) {
+		DisplayMetrics metrics = context.getResources().getDisplayMetrics();
+		return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dipValue, metrics);
+	}
+
+	public static int calculateInSampleSize(BitmapFactory.Options options, int reqWidth, int reqHeight) {
+		final int height = options.outHeight;
+		final int width = options.outWidth;
+		int inSampleSize = 1;
+
+		if (height > reqHeight || width > reqWidth) {
+			final int halfHeight = height / 2;
+			final int halfWidth = width / 2;
+			while ((halfHeight / inSampleSize) > reqHeight && (halfWidth / inSampleSize) > reqWidth)
+				inSampleSize *= 2;
+		}
+
+		return inSampleSize;
 	}
 }