+++ src/biz/bokhorst/xprivacy/XTelephonyManager.java
@@ -79,31 +79,40 @@ public class XTelephonyManager extends XHook {
 	// frameworks/base/telephony/java/android/telephony/TelephonyManager.java
 	// http://developer.android.com/reference/android/telephony/TelephonyManager.html
 
-	// public java.lang.String getDeviceId()
-	// public java.lang.String getDeviceSvn()
-	// public java.lang.String getSubscriberId()
-	// public java.lang.String getGroupIdLevel1()
-	// public java.lang.String getIccSerialNumber()
-	// public java.lang.String getLine1Number()
-	// public java.lang.String getLine1AlphaTag()
-	// public java.lang.String getMsisdn()
-	// public java.lang.String getVoiceMailNumber()
-	// public java.lang.String getCompleteVoiceMailNumber()
-	// public java.lang.String getVoiceMailAlphaTag()
-	// public java.lang.String getIsimImpi()
-	// public java.lang.String getIsimDomain()
-	// public java.lang.String[] getIsimImpu()
-	// http://grepcode.com/file/repository.grepcode.com/java/ext/com.google.android/android/4.4.4_r1/com/android/internal/telephony/PhoneSubInfo.java
+	// public String getDeviceId()
+	// public String getSubscriberId()
+	// public String getGroupIdLevel1()
+	// public String getIccSerialNumber()
+	// public String getImei()
+	// public String getLine1Number()
+	// public String getLine1AlphaTag()
+	// public String getMsisdn()
+	// public String getVoiceMailNumber()
+	// public String getVoiceMailAlphaTag()
+	// public String getCompleteVoiceMailNumber()
+	// public String getIsimImpi()
+	// public String getIsimDomain()
+	// public String[] getIsimImpu()
+	// public String getIsimIst()
+	// public String[] getIsimPcscf()
+	// http://grepcode.com/file/repository.grepcode.com/java/ext/com.google.android/android/5.0.0_r1/com/android/internal/telephony/PhoneSubInfo.java
 
 	// public void listen(java.lang.String pkg, IPhoneStateListener callback, int events, boolean notifyNow)
-	// http://grepcode.com/file/repository.grepcode.com/java/ext/com.google.android/android/4.4.4_r1/com/android/server/TelephonyRegistry.java
+	// http://grepcode.com/file/repository.grepcode.com/java/ext/com.google.android/android/5.0.0_r1/com/android/server/TelephonyRegistry.java
 
 	// public void enableLocationUpdates()
+	// public void enableLocationUpdatesForSubscriber(long subId)
 	// public void disableLocationUpdates()
-	// public java.util.List<android.telephony.CellInfo> getAllCellInfo()
+	// public void disableLocationUpdatesForSubscriber(long subId)
+	// public List<android.telephony.CellInfo> getAllCellInfo()
 	// public android.os.Bundle getCellLocation()
-	// public java.util.List<android.telephony.NeighboringCellInfo> getNeighboringCellInfo()
-	// http://grepcode.com/file/repository.grepcode.com/java/ext/com.google.android/android-apps/4.4.4_r1/com/android/phone/PhoneInterfaceManager.java#PhoneInterfaceManager
+	// public String getCdmaMdn(long subId)
+	// public String getCdmaMin(long subId)
+	// public String getLine1AlphaTagForDisplay(long subId)
+	// public String getLine1NumberForDisplay(long subId)
+	// public List<android.telephony.NeighboringCellInfo> getNeighboringCellInfo()
+	// http://grepcode.com/file/repository.grepcode.com/java/ext/com.google.android/android/5.0.0_r1/com/android/internal/telephony/ITelephony.java
+	// http://grepcode.com/file/repository.grepcode.com/java/ext/com.google.android/android-apps/5.0.0_r1/com/android/phone/PhoneInterfaceManager.java
 
 	// @formatter:on
 
@@ -128,11 +137,16 @@ public class XTelephonyManager extends XHook {
 		Srv_getMsisdn,
 		Srv_getSubscriberId,
 		Srv_getCompleteVoiceMailNumber, Srv_getVoiceMailNumber, Srv_getVoiceMailAlphaTag,
+		Srv_getImei, Srv_getIsimIst, Srv_getIsimPcscf,
 
 		Srv_listen,
 
 		Srv_enableLocationUpdates, Srv_disableLocationUpdates,
-		Srv_getAllCellInfo, Srv_getCellLocation, Srv_getNeighboringCellInfo
+		Srv_getAllCellInfo, Srv_getCellLocation, Srv_getNeighboringCellInfo,
+
+		Srv_enableLocationUpdatesForSubscriber, Srv_disableLocationUpdatesForSubscriber,
+		Srv_getCdmaMdn, Srv_getCdmaMin,
+		Srv_getLine1AlphaTagForDisplay, Srv_getLine1NumberForDisplay
 	};
 	// @formatter:on
 
@@ -190,6 +204,10 @@ public class XTelephonyManager extends XHook {
 			listHook.add(new XTelephonyManager(Methods.Srv_getVoiceMailAlphaTag, PrivacyManager.cPhone, Srv.SubInfo));
 			listHook.add(new XTelephonyManager(Methods.Srv_getVoiceMailNumber, PrivacyManager.cPhone, Srv.SubInfo));
 
+			listHook.add(new XTelephonyManager(Methods.Srv_getImei, PrivacyManager.cPhone, Srv.SubInfo));
+			listHook.add(new XTelephonyManager(Methods.Srv_getIsimIst, PrivacyManager.cPhone, Srv.SubInfo));
+			listHook.add(new XTelephonyManager(Methods.Srv_getIsimPcscf, PrivacyManager.cPhone, Srv.SubInfo));
+
 			listHook.add(new XTelephonyManager(Methods.Srv_listen, PrivacyManager.cLocation, Srv.Registry));
 			listHook.add(new XTelephonyManager(Methods.Srv_listen, PrivacyManager.cPhone, Srv.Registry));
 		}
@@ -204,6 +222,14 @@ public class XTelephonyManager extends XHook {
 			listHook.add(new XTelephonyManager(Methods.Srv_getAllCellInfo, PrivacyManager.cLocation, Srv.Phone));
 			listHook.add(new XTelephonyManager(Methods.Srv_getCellLocation, PrivacyManager.cLocation, Srv.Phone));
 			listHook.add(new XTelephonyManager(Methods.Srv_getNeighboringCellInfo, PrivacyManager.cLocation, Srv.Phone));
+
+			listHook.add(new XTelephonyManager(Methods.Srv_enableLocationUpdatesForSubscriber,
+					PrivacyManager.cLocation, Srv.Phone));
+			listHook.add(new XTelephonyManager(Methods.Srv_disableLocationUpdatesForSubscriber, null, Srv.Phone));
+			listHook.add(new XTelephonyManager(Methods.Srv_getCdmaMdn, PrivacyManager.cPhone, Srv.Phone));
+			listHook.add(new XTelephonyManager(Methods.Srv_getCdmaMin, PrivacyManager.cPhone, Srv.Phone));
+			listHook.add(new XTelephonyManager(Methods.Srv_getLine1AlphaTagForDisplay, PrivacyManager.cPhone, Srv.Phone));
+			listHook.add(new XTelephonyManager(Methods.Srv_getLine1NumberForDisplay, PrivacyManager.cPhone, Srv.Phone));
 		}
 		return listHook;
 	}
@@ -221,8 +247,14 @@ public class XTelephonyManager extends XHook {
 				param.setResult(null);
 			break;
 
+		case Srv_disableLocationUpdatesForSubscriber:
+			if (isRestricted(param, PrivacyManager.cLocation, "Srv_enableLocationUpdatesForSubscriber"))
+				param.setResult(null);
+			break;
+
 		case enableLocationUpdates:
 		case Srv_enableLocationUpdates:
+		case Srv_enableLocationUpdatesForSubscriber:
 			if (isRestricted(param))
 				param.setResult(null);
 			break;
@@ -308,6 +340,13 @@ public class XTelephonyManager extends XHook {
 		case Srv_getCompleteVoiceMailNumber:
 		case Srv_getVoiceMailNumber:
 		case Srv_getVoiceMailAlphaTag:
+		case Srv_getImei:
+		case Srv_getIsimIst:
+		case Srv_getIsimPcscf:
+		case Srv_getCdmaMdn:
+		case Srv_getCdmaMin:
+		case Srv_getLine1AlphaTagForDisplay:
+		case Srv_getLine1NumberForDisplay:
 			break;
 		}
 	}
@@ -321,6 +360,8 @@ public class XTelephonyManager extends XHook {
 		case enableLocationUpdates:
 		case Srv_disableLocationUpdates:
 		case Srv_enableLocationUpdates:
+		case Srv_disableLocationUpdatesForSubscriber:
+		case Srv_enableLocationUpdatesForSubscriber:
 			break;
 
 		case getAllCellInfo:
@@ -388,6 +429,7 @@ public class XTelephonyManager extends XHook {
 		case Srv_getIsimImpi:
 		case Srv_getIsimImpu:
 		case Srv_getSubscriberId:
+		case Srv_getImei:
 			if (param.getResult() != null)
 				if (isRestricted(param))
 					param.setResult(PrivacyManager.getDefacedProp(uid, mMethod.name().replace("Srv_", "")));
@@ -399,10 +441,22 @@ public class XTelephonyManager extends XHook {
 		case Srv_getCompleteVoiceMailNumber:
 		case Srv_getVoiceMailNumber:
 		case Srv_getVoiceMailAlphaTag:
+		case Srv_getLine1AlphaTagForDisplay:
+		case Srv_getLine1NumberForDisplay:
 			String srvPhoneNumber = (String) param.getResult();
 			if (srvPhoneNumber != null)
 				if (isRestrictedValue(param, srvPhoneNumber))
-					param.setResult(PrivacyManager.getDefacedProp(uid, mMethod.name().replace("Srv_", "")));
+					param.setResult(PrivacyManager.getDefacedProp(uid,
+							mMethod.name().replace("Srv_", "").replace("ForDisplay", "")));
+			break;
+
+		case Srv_getIsimIst:
+		case Srv_getIsimPcscf:
+		case Srv_getCdmaMdn:
+		case Srv_getCdmaMin:
+			if (param.getResult() != null)
+				if (isRestricted(param))
+					param.setResult(null);
 			break;
 		}
 	}